add as.list and merge method to poly_frame
`merge_fn` is now an attribute of poly_frame, it is used to merge the list of
data-frames together and can be accessed using `get_merge_fn`.

Tests to check that as.list and subsetting operators work fine on poly_frame
objects. Note that slicing a poly_frame returns a list, not a poly_frame.

poly_frame explicitly inherits from 'list' - this allows us to use subsetting
operators, and other list-methods like `length`.